GRIDEDIT-235: + Removed static specifiers in Operations.cpp by splitting Operations.cpp into Operations.cpp and Operations.hpp + Sorted all includes in all files + Remove unnecessary includes
