+++ include/MeshKernel/FlipEdges.hpp
@@ -26,6 +26,7 @@
 //------------------------------------------------------------------------------
 
 #pragma once
+#include <memory>
 
 namespace meshkernel
 {
