+++ include/MeshKernel/LandBoundaries.hpp
@@ -27,6 +27,7 @@
 
 #pragma once
 
+#include <memory>
 #include <vector>
 #include <MeshKernel/Entities.hpp>
