+++ src/FlipEdges.cpp
@@ -28,12 +28,12 @@
 #pragma once
 
 #include <vector>
-#include <MeshKernel/Operations.cpp>
 #include <MeshKernel/Entities.hpp>
-#include <MeshKernel/Mesh.hpp>
+#include <MeshKernel/Exceptions.hpp>
 #include <MeshKernel/FlipEdges.hpp>
 #include <MeshKernel/LandBoundaries.hpp>
-#include <MeshKernel/Exceptions.hpp>
+#include <MeshKernel/Mesh.hpp>
+#include <MeshKernel/Operations.hpp>
 
 meshkernel::FlipEdges::FlipEdges(std::shared_ptr<Mesh> mesh,
                                  std::shared_ptr<LandBoundaries> landBoundary,
