+++ src/TriangulationInterpolation.cpp
@@ -30,9 +30,9 @@
 #include <MeshKernel/TriangulationWrapper.hpp>
 
 #include <MeshKernel/Entities.hpp>
-#include <MeshKernel/Operations.cpp>
-#include <MeshKernel/SpatialTrees.hpp>
 #include <MeshKernel/Exceptions.hpp>
+#include <MeshKernel/Operations.hpp>
+#include <MeshKernel/SpatialTrees.hpp>
 
 meshkernel::TriangulationInterpolation::TriangulationInterpolation(const std::vector<Point>& m_locations,
                                                                    const std::vector<Sample>& samples,
