+++ tests/unit/CurvilinearGridFromSplinesTransfiniteTest.cpp
@@ -1,11 +1,9 @@
-#include <MeshKernel/Mesh.hpp>
-#include <MeshKernel/Entities.hpp>
-#include <MeshKernel/Polygons.hpp>
-#include <MeshKernel/Constants.hpp>
-#include <MeshKernel/Splines.hpp>
+#include <gtest/gtest.h>
 #include <MeshKernel/CurvilinearGrid.hpp>
 #include <MeshKernel/CurvilinearGridFromSplinesTransfinite.hpp>
-#include <gtest/gtest.h>
+#include <MeshKernel/Entities.hpp>
+#include <MeshKernel/Mesh.hpp>
+#include <MeshKernel/Splines.hpp>
 
 TEST(CurvilinearGridFromSplinesTransfinite, FourSplines)
 {
