+++ tests/unit/FlipEdgesTest.cpp
@@ -1,10 +1,10 @@
-#include <TestUtils/MakeMeshes.hpp>
-#include <MeshKernel/Entities.hpp>
-#include <MeshKernel/Polygons.hpp>
 #include <MeshKernel/Constants.hpp>
-#include <MeshKernel/Mesh.hpp>
-#include <MeshKernel/LandBoundaries.hpp>
+#include <MeshKernel/Entities.hpp>
 #include <MeshKernel/FlipEdges.hpp>
+#include <MeshKernel/LandBoundaries.hpp>
+#include <MeshKernel/Mesh.hpp>
+#include <MeshKernel/Polygons.hpp>
+#include <TestUtils/MakeMeshes.hpp>
 
 #include <gtest/gtest.h>
