+++ tests/unit/TriangleInterpolationTest.cpp
@@ -1,9 +1,9 @@
-#include <MeshKernel/Mesh.hpp>
 #include "Meshkernel/Entities.hpp"
-#include <MeshKernel/TriangulationInterpolation.hpp>
-#include <TestUtils/SampleFileReader.hpp>
-#include <TestUtils/MakeMeshes.hpp>
 #include <gtest/gtest.h>
+#include <MeshKernel/Mesh.hpp>
+#include <MeshKernel/TriangulationInterpolation.hpp>
+#include <TestUtils/MakeMeshes.hpp>
+#include <TestUtils/SampleFileReader.hpp>
 
 TEST(TriangleInterpolation, InterpolateOnNodes)
 {
